Object-Detection-model/requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
ultralytics==8.0.196
Pillow==10.0.1
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2